@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <StartupObject>BoolsAndCows.Program</StartupObject>
     <PlatformTarget>x64</PlatformTarget>
     <Nullable>enable</Nullable>
