@@ -1,7 +1,7 @@
-﻿global using ConsoleMenu;
+﻿global using BoolsAndCows.Enums;
+global using ConsoleMenu;
 global using System;
 global using System.Collections.Generic;
 global using System.Linq;
 global using System.Text;
 global using static BoolsAndCows.Data;
-global using static ConsoleMenu.Menu;
